requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
beautifulsoup4
scikit-learn
pandas
joblib
chromadb
sentence-transformers
openai
python-dotenv
pytest
fastapi
uvicorn[standard]
requests
beautifulsoup4
openai
faiss-cpu
numpy
python-dotenv
streamlit
pydantic
